Do not use cl-typep, which is broken for objects on Emacs 25
The issue we are avoiding is that `cl-typep' returns nil when it
should return t, because the VAL does in fact derive from the class
TYPE.

To avoid this issue we have to resort to using `CLASS--eieio-childp'
predicates, which--as the name suggests--we are not supposed to use.

`cl-typep' only appears to be broken when using byte-compiled code
and some other factor also appears to play a role.  I was not able
to reproduce this issue when using Borg instead of Package.  On
Emacs 26 `cl-typep' works as expected (though because it has no
doc-string, we don't know whether the expectation changed between
those two Emacs releases).

Also don't use `cl-etypecase' when objects are involved because
that uses `cl-typep'.

Fixes https://github.com/magit/magit/issues/3732.