transient--ensure-infix-command: Avoid false-positives
This function is supposed to generate the command for an anonymous
infix argument but until now it also generated a command for other
kinds of suffixes if the command was not defined or autoloaded.

Now this function detects that issue and raises an informative error
instead of defining a bogus command, which then caused a confusing
error later on.

Closes https://github.com/magit/magit/issues/3761.