transient--insert-suffix: Fix move of last child of a group
When adding a suffix using `transient-{insert,append}-suffix' then
an existing binding for the same command is removed, if any.  This
is necessary to either *move* the suffix binding or to replace it
with an alternative binding for the same command.  This can also
happen when the same form gets evaluated twice, which can happen
if it is autoloaded.

When the suffix is the last child of a group, then this was broken.
This is the same bug as was fixed with 0ad64ddc5db466fb3bcac67b931,
but in a different function.

Fixes #20.