Declare that prefix and suffix commands are for interactive use only
I intended to do this for a long time but never got around to do it.

In all of Magit there were only four cases where a suffix command
was called non-interactively.  One of them was a bug, another a very
special case where the command calls itself non-interactively (now
inside `with-no-warnings') and only two cases where we now have to
add a tiny bit of duplication to avoid the non-interactive calls.

I think that means calling suffix commands non-interactively is not
something that is useful often in general.  If those very few cases
were calling non-interactively is legitimate `with-no-warning' or
maybe an explicit "(put SUFFIX 'interactive-only nil)" after the
suffix definition should be good enough.

For prefix commands I think it is simply always wrong to call them
non-interactively.