transient-set-level: Limit when it is called
Only arrange for `transient-set-level' to be called if
`transient-suffix-object' actually returns an object.  This prevents
users from setting the level of commands shared between all suffixes.
Setting the level for such a command wouldn't have an effect and some
of them are useful while setting the levels for other commands, e.g.
the scroll commands.

Closes #29.