transient--delete-window: Protect against misconfiguration
Transient expects the popup buffer to be displayed in a new window.
If that is not the case, then it won't work properly and this commit
does not do anything to make it work properly anyway (because that
is not possible).

Regardless, the user should still be able to exit the transient state
even when things are broken like that, but doing so failed if the
transient popup was displayed in the only window because we ended up
trying to delete window, which is not allowed.  Continue to try to
kill the transient window, but if that fails, then continue with the
rest of the exit procedure.

Re #34 #44.