No longer allow characters as pseudo suffixes
The only potentially useful character were ?\n and ?\s but "" has the
same effect as these two.  Unlike integers two `equal' strings are not
also `eq', which means that attempting to remove an integer by
position (to be implemented in subsequent commits) would always remove
the first `eq' integer instead of the specified one.