Remove conflicting suffix when inserting new suffix
When inserting a suffix and another suffix is bound to same key, then
remove the other suffix.  I.e. more or less do what `define-key' would
do.

I believe doing that was already the intention before but instead of
actually doing that another binding for the suffix that is being was
removed.  So `transient-insert-suffix' could be used to move where in
the popup a suffix appeared, which may be useful in some odd cases,
but isn't what this function is supposed to do.