Split new transient-get-value from transient-args
Going forward `transient-args' is only intended for use by regular
suffix commands that consume the arguments.

The new `transient-get-value' is intended primarily for indirect use
by special suffix commands that set or save the value.  More precisely
by the methods `transient-set-value' and `transient-save-value', which
are used by the commands `transient-set' and `transient-save' (as well
as by the internal `transient--history-push').