transient--insert-suffix: Fix use of binding defined in object
The key binding of SUFFIX may be defined in the suffix object or in
the suffix spec.  Previously we only considered the latter and if that
binding was nil, then we ended up treating nil as the binding, which
then resulted in the removal of the suffix specified by LOC from the
PREFIX.

Fixes #58.