transient-init-value: Always prefer saved over default value
The previous implementation would have worked if only `slot-boundp'
returned nil for slots that have a default value that is *not*
shadowed by an explicitly set value.  But that is not so; the slot
is considered bound in that case too.

So we need two slots -- one for the default and one for the actual
value.  We keep using `value' for the actual value because that is
used much more often.  Only when implementing a `transient-init-value'
method should one have to deal with the new `default-value' slot.
Note that `value' now has no initarg and that `default-value's initarg
is `:value'; but luckily that is an implementation detail that users
of the `define-transient-command' macro don't have to know about that.

Closes #68.