manual: Remove some trigger words from LocalWords
Without removing these words, Ispell complains like so:

ispell--run-on-word: Ispell and its process have different character
maps: (* Error: The word "texinfo+" is invalid. The character '+'
(U+2B) may not appear at the end of a word. Error: The word
"pre-command" is invalid. The character '-' (U+2D) may not appear in
the middle of a word. Error: The word "Magit-Popup" is invalid. The
character '-' (U+2D) may not appear in the middle of a word.)