transient-setup: Handle edge-case
When the minibuffer is active on behalf of a suffix command, then we
cannot completely prevent the user from invoking commands that they
should not invoke at this time (i.e. they can use recursive editing,
which does not play well with transient prefixes) because we have to
allow arbitrary commands as we cannot know whether any such command
is something that aids during minibuffer completion.

When the user takes the opportunity to invoke another prefix at this
time, then this leads to events from our minibuffer and setup hooks
being firing in an unintended order and that leads to Emacs entering
an inconsistent transient state.

Since we wouldn't want the user to move away from the minibuffer even
if doing so did not lead to this issue, we simply detect and disallow
entering a transient prefix after having done that.

Closes the variant of #64 that was reported by @bricewge.