transient--emergency-exit: Always signal the error
If some code is wrapped inside `transient--with-emergency-exit' then
it is enough to call `error' but the code that raises the error might
be far away from the code that uses the macro, so we might end up
needlessly calling `transient--emergency-exit' in the local context.

,----
| (transient--with-emergency-exit
|   ;; The bove and below parts likely located in different functions.
|   (when (something-bad-p)
|     (transient--emergency-exit)
|     (error "Something bad has happened")))
`----

The explicit call to `transient--emergency-exit' sets
`transient--prefix' to nil and the second call that is triggered by
`error' used to fail to raise the error again because it used to not
do anything if `transient--prefix' is nil.