transient--insert-suffix: Allow same key with different predicates
Previously it was not possible to add two bindings that use the
same key but different predicates using `transient-insert-suffix'.
Such only superficially conflicting bindings had to be added when
the prefix was defined using `transient-define-prefix'.

Now we use a heuristic to allow `transient-insert-suffix' to add
such bindings as well.

If both bindings define a predicate and these predicates are not
equal, then the two bindings are assumed to be distinct.  If on
the other hand at least one of the bindings lacks a predicate or
the predicates are equal, then the new binding is considered a
replacement for the old binding, as before.

Closes #106.