Allow invoking arbitrary prefixes as non-suffixes
When a prefix is explicitly added as a suffix of another "outer"
prefix, then the `transient--do-replace' pre-command function is
added to the predicate map by `transient--make-predicate-map',
which is aware that this is a sub-prefix.

When the "outer" prefix allows arbitrary non-suffix commands to
be invoked while it is active by setting `transient--do-stay' or
`transient--do-call', then there is no opportunity to configure
another predicate for commands that happen to be prefixes.

This results in a pre-command being called that is not appropriate
for a sub-prefix command because it fails to clean up after the
"outer" prefix command.  In particular it fails to remove to active
transient keymaps before adding new keymaps of the new prefix.

That causes them to stick around forever.  The redisplay keymap ends
up consume certain key events, such as "C-x", without unreading them
again so that the real bindings can be looked up.

We deal with this by teaching the code that sets up a prefix to check
whether a transient keymap (`transient--transient-map') is still in
effect and if so to take this as a sign that the above is what has
happened and to clean up accordingly.

This isn't considered a kludge but proper behavior.  Another place
where this could be detected is `transient--pre-command', by checking
if (get this-command 'transient--prefix) is non-nil and then pivot to
`transient--do-regexp'.

diff --git a/lisp/transient.el b/lisp/transient.el
@@ -1865,10 +1865,15 @@ (defun transient--pre-command ()
     (setq transient--exitp nil)
     (when (eq (if-let ((fn (transient--get-predicate-for
                             this-original-command)))
-                  (let ((action (funcall fn)))
-                    (when (eq action transient--exit)
-                      (setq transient--exitp (or transient--exitp t)))
-                    action)
+                  (progn
+                    (when (and (get this-command 'transient--prefix)
+                               (memq fn '(transient--do-stay
+                                          transient--do-call)))
+                      (setq fn 'transient--do-replace))
+                    (let ((action (funcall fn)))
+                      (when (eq action transient--exit)
+                        (setq transient--exitp (or transient--exitp t)))
+                      action))
                 (if (let ((keys (this-command-keys-vector)))
                       (eq (aref keys (1- (length keys))) ?\C-g))
                     (setq this-command 'transient-noop)