Allow using transients from inside active minibuffer
It was already possible to invoke a transient prefix command while
the minibuffer is active.  However when an infix command is invoked
from that prefix, which itself uses the minibuffer to read input,
then the keymaps and hooks that have to be suspended while reading
in the minibuffer were not actually suspended.

`transient--minibuffer-setup' and `transient--minibuffer-exit'
actively avoided doing anything if `minibuffer-depth' is greater
than 1, which here it is.  I was unable to figure out why they did
that and have simply removed that check.

Additionally we have to let-bind `enable-recursive-minibuffers' to
t, or else it would be possible for a user to invoke the transient
but then to be disallowed to fully use that transient.

Finally, exiting the transient caused the minibuffer window to be
deselected in this case, forcing us to actively re-select that window.

Closes #109.