transient-read-directory: Pass read value through file-local-name
Same for `transient-read-existing-directory'.

I am confident that most, if not all, existing and future callers
expect a local name (i.e. all callers that pass arguments to some
child process, and that is the original use-case after all). If it
turns out that some callers need the remote part, then we can add
new variants for that.

Closes https://github.com/magit/magit/issues/4322.