Allow invoking arbitrary prefixes as non-suffixes
When a prefix is explicitly added as a suffix of another "outer"
prefix, then the `transient--do-replace' pre-command function is
added to the predicate map by `transient--make-predicate-map',
which is aware that this is a sub-prefix.

When the "outer" prefix allows arbitrary non-suffix commands to
be invoked while it is active, by setting `transient-non-suffix'
to `transient--do-stay' or `transient--do-call', then there is no
opportunity to configure another predicate for commands that happen
to be prefixes.

This results in a pre-command being called that is not appropriate
for a sub-prefix command because it fails to clean up after the
"outer" prefix command.  In particular it fails to remove the active
transient keymaps before adding new keymaps of the new prefix.

That causes them to stick around forever.  The redisplay keymap ends
up consume certain key events, such as "C-x", without unreading them
again so that the real bindings can be looked up.

We deal with this by teaching the code that sets up a prefix to check
whether a transient keymap (`transient--transient-map') is still in
effect and if so to take this as a sign that the above is what has
happened and to clean up accordingly.