Quote value of slots' :initform when defining classes
The behavior of `:initform' changes in Emacs 28:

> Change the implementation of `:initform` to better match the CLOS
> semantics, while preserving the EIEIO semantics, but warn when
> encountering cases where the two diverge.

As a result we have to quote the name of two functions.

We also have to obfuscate using the value of a variable in another
instance because the transitional code would otherwise not be able
to tell whether this is intentional or assumes the old semantics.