Enable section movement
This doesn't actually implement any section movement commands.
It merely makes it possible to implement such commands without
having to modify or advice transient.

See https://github.com/magit/magit/issues/4438.

This stops placing the cursor on the first suffix.  Instead the
cursor is now left at the beginning of the buffer, on the first
section heading.  This is beneficial for blind users who rely
on a screen reader.  The old hack that moved the cursor did not
only do that initially, it actually prevented from ever moving
before the first suffix.