transient--read-number-N: Return empty string instead of nil
This affects `transient-read-number-N+' and `transient-read-number-N',
which are only intended for use as `:reader' of infix commands and are
called by `transient-infix-read' which understand how to treat nil and
the empty string depending on the values of other slots.

For empty input return the empty string instead of nil, informing
`transient-infix-read' that what is being returned is the empty value
of the argument and not the instruction to altogether turn off the
argument.  This makes it trivial to support options that can be off,
on, or have a numeric value:

 (transient-define-argument magit-diff:-C ()
   :description "Detect copies"
   :class 'transient-option
   :argument "-C"
+  :allow-empty t
   :reader 'transient-read-number-N+)