Export variables for transient non-infix suffixes
Unlike infixes, which change the value, non-infix suffixes are
supposed to use the value.  Just because they are made transient,
that doesn't mean they should fall back to the default value and
have no access to other state variables that need to be exported.

It was clearly unintentional to use the non-exporting pre-command
for non-infix suffixes when the value of `transient' is just t.

Closes #151.