More flexible display of transient buffer
Display buffer after populating it and set additional display actions
and window parameters using the display alist instead of hardcoding
them.  (`set-window-hscroll' should be unnecessary so drop that.)

Use a wrapper function around `fit-window-to-buffer' so that users can
override it (but do not go as far as to provide a hook).

Continue to call that function explicitly because the `window-height'
display action is ignored when reusing an existing window, and while
we don't encourage doing that, we still want to support it.  If an
existing window is repurposed, then avoid shrinking it.