Fix handling of sub-prefix command that use the minibuffer
When a suffix command used the minibuffer, then we have to suspend the
transient's hooks and keymaps while doing so and `post-command-hook'
is called twice, once before entering the minibuffer and then again
when the command is actually done.  On the first call we delay work
until either the second call or until the minibuffer is exited
abnormally.

This also applies if the suffix command is itself a prefix command.

A prefix command sets up the transient by calling `transient-setup' in
its body.  If that command uses the minibuffer then it does so before
calling `transient-setup'.

If the minibuffer is aborted, then the transient state is cleaned up
using a function that `transient--delay-post-command' added to
`minibuffer-exit-hook'.  This case works reasonably but can be
improved (see next commit).

If the minibuffer is exited normally, then the transient state is
adjusted by a function that `transient--delay-post-command' put on
`transient--post-command-hook'.

But before that happens `transient-setup' is called.  That puts the
normal function on `post-command-hook'.  Then the sub-prefix command
finishes and that hook function is called, which calls
`transient--post-exit'.

But then the hook function that was put on
`transient--post-command-hook' is also called and it also calls
`transient--post-exit'.  If that second call was a noop that wouldn't
be a huge deal but it is not.  Now it calls `transient--stack-pop',
causing a return to the outer prefix.

Preventing that is easy.  On the pre-minibuffer call to
`transient--post-command' we can detect that the suffix is a
sub-prefix and in that case we only delay work until the minibuffer is
aborted (if it is aborted) but not until the `post-command-hook' is
called when the command is actually done.

Closes #187.