Do not let-bind overriding-terminal-local-map to nil
Previously we let-bound `overriding-terminal-local-map' to nil in
`transient-infix-read', which is not only unnecessary (because we
temporarily remove the transient map through other means).  When
the debugger is active and an error is signaled while the binding
is active, this resulted in the transient map being restored after
exiting the debugger.