Only use this-original-command when necessary
It is necessary for `transient-help' and `transient-set-level'.
Invoking one of these commands enters a special mode.  The next
command is always replaced with the command that entered the
special mode, which now behaves differently: it show the help
resp. sets the level for the `this-original-command'.

Other uses of `this-original-command' likely were "inspired" by
these legitimate uses.