Use this-original-command again
[1: 06697f75] removed some uses of `this-original-command' that were
deemed unnecessary.  In some case that turned out to be a mistake.
This commit reverts those changes and adds comments to prevent the
same mistake in the future.

In one case the logic is improved to reset `this-command' to
`this-original-command' instead of just reverting to using the latter.
Doing it that way fixes an older bug that prevented the remapping from
one non-suffix to another.

Closes #195.

1: 2022-03-25 06697f75ace79936e12709cfbd75547b7016b733
   Only use this-original-command when necessary