transient--show: Also hide the header-line
Some Emacs users move the content of the mode-line to the
header-line. This is also what nano-modeline [1] does by
default. `transient--show` hides the mode-line and the tab-line but
not the header-line. As a result, the last line of transient buffers
are invisible for these users. Hiding the header-line fixes the issue.

https://github.com/rougier/nano-modeline