Support use of an infix argument following a prefix argument
Using a prefix argument activates `universal-argument--mode' whose
keymap, `universal-argument-map', binds "-" to `negative-argument'.
That binding conflicts with the prefix key used for most infix
arguments.

Counter that by "remapping" `negative-argument' to `transient-update',
using a dedicated pre-command, but only if `negative-argument' was
invoked using "-".  "C--", own binding for that command, remains
functional.

When remapped from `negative-argument', `transient-update' has to
preserve the prefix argument for the next command.  This cannot happen
earlier in the pre-command, because Emacs consumes `prefix-arg' in
between `pre-command-hook' and the actual invocation of the command
itself.

`prefix-arg' has to be set directly because the function normally used
to do so, `prefix-command-preserve-state', sets `this-command' to
`last-command', which would mess with `transient--post-command'.
Otherwise if the sequence of commands is (1) transient prefix command,
(2) prefix argument, (3) infix argument, then `transient--post-command'
would then act as if the prefix and infix arguments were never invoked.