Fix invoking the suffix the cursor is on in the transient buffer
In [1: bb056e71] we changed how suffix commands are invoked when a
button is activated because the old way resulted in the transient
buffer being current when the suffix command is invoked, which messes
things up.

Clicking on a button in the transient buffer now only temporarily
makes the transient buffer current.  Other bindings had to be moved
from `transient-button-map' to `transient-popup-navigation'.  The
binding for invoking the suffix "under the cursor" was overlooked.

Closes #197.

1: 2022-03-25 bb056e7156b3d88f42770ec55e1a7447a95aca96
   Invoke suffix commands directly when a button is pushed