Change key binding for invoking the current suffix
The binding is in `transient-popup-navigation-map', which means that
it is always available (provided popup navigation is enabled of).

Some transient prefixes allow arbitrary non-suffix commands to be
invoked, unless of course a binding of a suffix command conflicts.
Whatever non-suffix command "RET" would invoked, that's very likely
one that we should not shadow.  So bind "M-RET", which in most cases
should be bound to a command that is at least less frequently used
than whatever "RET" is bound to.