transient--delay-post-command: Handle execute-extended-command
When using `execute-extended-command' to call another command that
also uses the minibuffer, then this hook function is never called
with `this-command' being `execute-extended-command' and the event
being non-nil, so this failed to resume the transient state.

`execute-extended-command' sets `this-command' and `real-this-command'
so we need to rely on a different signal to determine when to resume.
This function also changes the value returned by `this-command-keys'
in predictable fashion, so we can rely on that.

Not that we would want to do that anyway, but note that editing
`execute-extended-command' so that it does not set `this-command' and
`real-this-command' does not work.  I was unable to determine where
that happens, but something else is setting these variable on behalf
of this command.  Using a (renamed) copy of this command with only the
code that sets these variables removed, however succeeds.