transient--describe-function: Handle renamed help buffer
We have to make sure that the help buffer becomes the current buffer
so that typing "q" resumes the transient as advertised.  When a hook
is used to rename "the" help buffer, then the previous approach
resulted in an error because we reasonably, but wrongly, assumed
that we would always be able to find the buffer using `help-buffer'.

Closes #208.