Turn suffix specifications into code instead of data
The macro `transient-define-prefix' manipulates its GROUP arguments,
but until now it expanded to new vectors using vector syntax, i.e.,
data.  An effort was made to evaluate lambda expressions, but that did
not really work.

Now the GROUP arguments are expanded into code, which evaluates to new
vectors.  This way lambda expressions are automatically evaluated and
byte-compiled, but we now have to take care to quote certain lists and
symbols.

Functions such as `transient-insert-suffix' take group and suffix
specs of the same form as `transient-define-prefix' but because there
is no macro expansion step, they have to `eval' the result of parsing
these specifications.

Add a new macro `transient-define-groups', which can be used to define
suffix groups that are shared between multiple prefix commands, but
don't use it for `transient-common-commands' to avoid having to make
a bunch of functions available at compile time.