transient-prompt(lisp-variable): Fallback to generic prompt
Teach `transient-prompt' to fall back to a generic prompt for
`transient-lisp-variable'.  If a prompt is explicitly provided, then
continue to use that.

(The method for `transient-infix', which is currently being used for
all other types, already falls back to a generic prompt in most cases,
but may in some cases use a prompt that warns that a prompt should be
given explicitly.)

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>