readers :allow-empty return original value
Setting value to nil when :allow-empty was non-nil caused the whole
expression to return nil to the transient-set method.

This change stores the original value so that it can be shunted back
into transient-set in the empty string case.

By making the change in transient-infix-read, the logic remains
centralized, keeping transient-infix-set usually very simple.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>