transient-help: Fix regression
That is one of those surprise bug fixes.  While investigating why
`transient-help' no longer worked, I noticed that it needlessly used
`called-interactively-p'.  To my surprise, changing that, also fixed
the issue I was about to work on.  oO

As I recently learned, `called-interactively-p' should be avoided if
at all possible.  It is inherently a hack and not guaranteed to always
work correctly.  This commit fixes a regression that was introduced
with [1: 51c68c8], where we started to advise `this-command'; so it
appears the use of around advice is one of the things that can cause
this function to misbehave.

1: 2023-02-08 51c68c87cc66322db2ee3d85b6f5fc8b941a2ed6
   Use around advice to deal with skipped post-command-hook