Provide two implementations of transient--wrap-command
Starting with Emacs 30 the `interactive' form of a command can be
advised without any trickery and the new `transient--wrap-command-30'
takes advantage of that.  Additionally it uses `letrec'.

On Emacs 29 and before we have to continue to use the old approach.
I was planning to stick to this approach until I can drop support for
these releases, but then Stefan changed `transient--wrap-command' on
Emacs' "master" branch to take advantage of the improved support for
advising commands.

To avoid further complicating the process of merging Transient into
Emacs (I already need a playbook to do it), I have decided to provide
two variants of this function.  I actually have implemented just that
a long time ago, but decided against installing it, because, while
quite ugly, the backward compatible implementation continues to work
fine, even in Emacs 30.  Now I feel, I have no choice but to do it
anyway.