transient--post-command: Redisplay after universal argument
We want to redisplay the transient buffer after an universal argument
so that suffix descriptions can reflect whether that is the case or
not.

If `this-command' is the current transient prefix, then we have
already taken care of updating the transient buffer elsewhere, so
we do not update here.

However, because `prefix-command-preserve-state' changes the value the
values of `this-command' and `real-this-command' to the previous
command, in our case potentially the transient prefix argument itself,
we often ended up not updating transient buffer after a universal
argument.

We can fix that by updating if `prefix-arg' is non-nil.

We cannot use `current-prefix-arg' for that purpose because it is set
too late (in `command-execute').

If that gets fixed, then we likely still would not be able to rely on
`current-prefix-arg' and `prefix-command-preserve-state-hook' would
have to be used record that a universal argument is in effect.