transient-update: Preserve universal argument
Since [1: ed2febd0] we have already done so, iff we remapped from
`negative-argument' to this command.

Now we always preserve the universal argument.  This is necessary
to pass such argument to suffix commands, which are bound to key
sequences longer than one event.  In such cases there is an
additional transient keymap, which binds the prefix key itself to
`transient-update'.  Using that binding, causes an update of the
transient buffer, before the key is unread so that it can be looked
up in the transient keymap with the real suffix command bindings.

Like any other command `transient-update' consumes the universal
argument, and we have to set `prefix-arg' again, as if it were
itself an universal argument command.

1: 2022-04-24 ed2febd0056932689da00414af9db0260ea08ead
   Support use of an infix argument following a prefix argument