Revert "transient-{set,save,reset}: Stay transient"
This reverts commit 361d8ac61bf75e027173f84a1013d87364a4ec3e.

When transient--do-stay is used, transient does not export variables.
This causes it to save the command without any of its arguments present.

For example, say one wants to only pass "--decorate" to magit-log. One
would open the log menu and set the "--decorate" option, and then save
it via C-x C-s. With do-call, the .emacs.d/.local/etc/transient/values
file looks like:

    $ cat values
    ((magit-log:magit-log-mode "--decorate"))

With do-stay, it looks like:

    $ cat values
    ((magit-log:magit-log-mode))

The result of this bug is that whenever you save any set of options, the
next time around /all/ of your options will be unselected.