Revert "transient-{set,save,reset}: Stay transient"
This reverts commit 361d8ac61bf75e027173f84a1013d87364a4ec3e.

The goal of that commit was to ensure that the transient stays
active when invoking one of these commands.

However both of these pre-commands are supposed to do that, so while
moving from one to the other, did seem to address that issue,
something else must be going on, which will have to be investigated
(including why `transient--do-call' only misbehaves for these suffix
commands in combination of certain prefix commands).

More importantly though, this commit has to be reverted because it
introduced a regression, which is worse than the minor defect it fixed.
These commands depend on variables being exported.  The only intended
difference between the two pre-commands is that `transient--do-call'
exports variables, so switching to `transient--do-stay' made no sense.

Closes #257.