transient-{set,save}-and-exit: New commands
Actually they are just aliases.  The separate entries
in `transient-predicate-map' is what makes the different.

These are intended for use in, e.g., `magit-diff-refresh'
and `magit-log-refresh', which use specialized classes whose
`transient-set-value' and `transient-save-value' additionally refresh
the buffer containing the diff/log.  I.e., there this commands serve
as variants of `magit-refresh', and all these commands are expected to
exit the transient.

Previously these transients instead bound `transient-set-value'
to "s" and `transient-save-value' to "w", and explicitly set their
`transient' slot to nil, but that was surprising because that also
causes the regular "C-x s" and "C-x C-s" bindings to behave that
way.