transient-group: Always bind PAD-KEYS slot
By not making a distinction between padding being disabled and padding
not being specified, subgroups loose the ability to disable padding
that was enabled in a parent group.

But that doesn't seem useful anyway; if one sibling group should not
use padding, while the others do, then it seems more reasonable to
just enable the padding individually.

This change also enables the next commit.